% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{adjust_density}
\alias{adjust_density}
\title{adjusts existing density object representing the mosaic fraction curve for each ID-sample combination}
\usage{
adjust_density(nested_data_filtered, nested_data, by_var)
}
\arguments{
\item{nested_data_filtered}{nested_data object with an existing density column}

\item{nested_data}{nested_data object with an existing density column}
}
\value{
nested_data object with density column with adjusted counts
}
\description{
adjusts existing density object representing the mosaic fraction curve for each ID-sample combination
}
