% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{annotate_mut_node}
\alias{annotate_mut_node}
\title{assigns node from gr to each unique mutation in raw data}
\usage{
annotate_mut_node(raw, gr)
}
\arguments{
\item{raw}{raw object with ID, sequence, and tissue columns present in data}

\item{gr}{tree from which nodes are to be assigned}
}
\value{
raw object with addition column gr_node in data
}
\description{
assigns node from gr to each unique mutation in raw data
}
