% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{annotate_parental_spacers}
\alias{annotate_parental_spacers}
\title{identifies mutated alleles based on existence within parental set of spacers}
\usage{
annotate_parental_spacers(raw, parent_table)
}
\arguments{
\item{raw}{parent_table holding raw data for each allele, including the following columns: ID, sequence}

\item{parent_table}{parent_table holding founder set of IDs and spacer sequences, including the following columns: ID, sequence}
}
\value{
raw holding raw data for each allele, including the following added column: mutated
}
\description{
identifies mutated alleles based on existence within parental set of spacers
}
