% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{append_density}
\alias{append_density}
\title{appends a density object representing the mosaic fraction curve for each ID-sample combination}
\usage{
append_density(nested_data, breaks)
}
\arguments{
\item{nested_data}{nested_data object}

\item{breaks}{vector of values to define desired bins to separate mosaic fraction values into}
}
\value{
nested_data object with density column added
}
\description{
appends a density object representing the mosaic fraction curve for each ID-sample combination
}
