% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{append_metadata}
\alias{append_metadata}
\title{appends any desired metadata to raw data based on specified columns}
\usage{
append_metadata(raw, metadata, join_by = NULL, cols = NULL)
}
\arguments{
\item{raw}{raw holding raw data for each allele, including the following columns: cols}

\item{metadata}{metadata with columns of interest to be appended to raw data, including the following columns: cols}

\item{join_by}{vector of column names to join (left join) metadata by}

\item{cols}{columns of interest in metadata to append to raw data}
}
\value{
raw holding raw data for each allele with all additional columns in metadata
}
\description{
appends any desired metadata to raw data based on specified columns
}
