% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_util.R
\name{assign_mut_node}
\alias{assign_mut_node}
\title{assigns each unique mutation in list of mutation (ID sequence combination) tip combinations with node in gr}
\usage{
assign_mut_node(mut_tip_list, gr)
}
\arguments{
\item{mut_tip_list}{list of mutation tip combinations}

\item{gr}{tree from which nodes are to be assigned to mutation tip combinations}
}
\value{
table of mutation tip combinations and assigned nodes
}
\description{
assigns each unique mutation in list of mutation (ID sequence combination) tip combinations with node in gr
}
