% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{calculate_density}
\alias{calculate_density}
\title{calculates mosaic fraction density curve given mosaic fraction data and desired bins}
\usage{
calculate_density(mf_vec, breaks)
}
\arguments{
\item{mf_vec}{vector of mosaic fraction values to create density curve to represent the distribution of mosaic fractions}

\item{breaks}{vector of values to define desired bins to separate mosaic fraction values into}
}
\value{
mf_density representing density curve
}
\description{
calculates mosaic fraction density curve given mosaic fraction data and desired bins
}
