% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{calculate_distribution}
\alias{calculate_distribution}
\title{#' calculates mosaic fraction density curve given mosaic fraction data and existing calculated density
#'
#' @param mf_vec vector of mosaic fraction values to create density curve to represent the distribution of mosaic fractions
#' @param mf_density mf_density object representing previously calculated density curve
#' @return mf_density representing new density curve with old cumulative mosaic fraction information
recalculate_density <- function(mf_vec, mf_density) {
  counts = calculate_density(mf_vec, mf_density$breaks)$table
  mf_density$table$count = counts$count
  mf_density
}
calculates mosaic fraction distribution given mosaic fraction data and desired bins}
\usage{
calculate_distribution(mf_vec, breaks)
}
\arguments{
\item{mf_vec}{vector of mosaic fraction values to create density curve to represent the distribution of mosaic fractions across bins}

\item{breaks}{vector of values to define desired bins to separate mosaic fraction values into}
}
\value{
mf_distribution representing distribution of mosaic fractions across bins
}
\description{
#' calculates mosaic fraction density curve given mosaic fraction data and existing calculated density
#'
#' @param mf_vec vector of mosaic fraction values to create density curve to represent the distribution of mosaic fractions
#' @param mf_density mf_density object representing previously calculated density curve
#' @return mf_density representing new density curve with old cumulative mosaic fraction information
recalculate_density <- function(mf_vec, mf_density) {
  counts = calculate_density(mf_vec, mf_density$breaks)$table
  mf_density$table$count = counts$count
  mf_density
}
calculates mosaic fraction distribution given mosaic fraction data and desired bins
}
