% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{compute_mosaic_fraction}
\alias{compute_mosaic_fraction}
\title{calculates mosaic fraction for each mutated allele in each sample}
\usage{
compute_mosaic_fraction(raw)
}
\arguments{
\item{raw}{raw holding raw data for each allele, including the following columns: sample, ID, sequence, reads}
}
\value{
raw holding raw data for each allele, including the following added columns: mosaic_fraction and log2MF
}
\description{
calculates mosaic fraction for each mutated allele in each sample
}
