% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{compute_mutation_information}
\alias{compute_mutation_information}
\title{calculates number of unique mutations and mutated fraction for each ID-sample combination}
\usage{
compute_mutation_information(nested_data)
}
\arguments{
\item{nested_data}{nested_data object}
}
\value{
nested_data object, including the following added columns: num_mut, mutated_fraction
}
\description{
calculates number of unique mutations and mutated fraction for each ID-sample combination
}
