% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{filter_recurring_spacers}
\alias{filter_recurring_spacers}
\title{filters ID spacer combinations based on indelphi calculated emergence probability}
\usage{
filter_recurring_spacers(
  raw,
  emergence_table,
  prob_cutoff,
  filter_missing = TRUE
)
}
\arguments{
\item{raw}{raw holding raw data for each allele, including the following columns: ID, sequence}

\item{emergence_table}{emergence_table holding founder set of IDs and spacer sequences, including the following columns: ID, sequence, probability}

\item{prob_cutoff}{double probability cutoff to subset mutations by}

\item{include_missing}{boolean indicating whether alleles not found in emergence_table should be filtered out}
}
\value{
raw holding raw data for each allele, including the following added column: probability
}
\description{
filters ID spacer combinations based on indelphi calculated emergence probability
}
