% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{filter_unmutated}
\alias{filter_unmutated}
\title{filters unmutated alleles out of raw data for mosaic fraction analysis}
\usage{
filter_unmutated(raw)
}
\arguments{
\item{raw}{raw holding raw data for each allele, including the following columns: mutated}
}
\value{
raw holding both filtered and unfiltered versions of raw data for each allele
}
\description{
filters unmutated alleles out of raw data for mosaic fraction analysis
}
