% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{id_filter}
\alias{id_filter}
\title{filters IDs based on existence within founder set of ID sequence combinations}
\usage{
id_filter(raw, parent_table)
}
\arguments{
\item{raw}{raw holding raw data for each allele, including the following columns: ID, sequence}

\item{parent_table}{parent_table holding founder set of IDs and spacer sequences, including the following columns: ID, sequence}
}
\value{
raw holding raw data for each allele
}
\description{
filters IDs based on existence within founder set of ID sequence combinations
}
