% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_util.R
\name{ks_test}
\alias{ks_test}
\title{creates ks_test_result object with results of two sample KS test between two mosaic fraction distributions}
\usage{
ks_test(mf_distribtion1, mf_distribution2)
}
\arguments{
\item{mf_distribution2}{mf_distribution to be compared to mf_distribution1}

\item{mf_distribution1}{mf_distribution}
}
\value{
ks_test_result object with results of two sample KS test
}
\description{
creates ks_test_result object with results of two sample KS test between two mosaic fraction distributions
}
