% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{load_parent}
\alias{load_parent}
\title{loads parent data into a parent_table object}
\usage{
load_parent(filepath)
}
\arguments{
\item{filepath}{string path to location of file}
}
\value{
parent_table holding parent data
}
\description{
loads parent data into a parent_table object
}
