% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{load_samples}
\alias{load_samples}
\title{loads sample data into a raw object}
\usage{
load_samples(file_paths)
}
\arguments{
\item{file_paths}{string path to location of files}
}
\value{
raw holding raw data for each allele
}
\description{
loads sample data into a raw object
}
