% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{nest_raw}
\alias{nest_raw}
\title{groups raw data by ID and sample}
\usage{
nest_raw(raw, by_var)
}
\arguments{
\item{raw}{raw holding raw data for each allele, including the following columns: sample, ID}
}
\value{
nested_data holding raw data for each allele in both raw and nested forms, nested by ID and sample with all additional data in dataframe data
}
\description{
groups raw data by ID and sample
}
