% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{normalize_density}
\alias{normalize_density}
\title{calculates normalized mosaic fraction density curve given previously calculated density}
\usage{
normalize_density(mf_density, apply_cum_adjustment = TRUE)
}
\arguments{
\item{mf_density}{mf_density object with latest_count column}
}
\value{
mf_density representing same density with norm_count column
}
\description{
calculates normalized mosaic fraction density curve given previously calculated density
}
