% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{normalize_density_list}
\alias{normalize_density_list}
\title{calculates normalized mosaic fraction density curves given previously calculated densities}
\usage{
normalize_density_list(nested_data, apply_cum_adjustment = TRUE)
}
\arguments{
\item{nested_data}{nested_data object with adjusted density column}
}
\value{
nested_data with normalized counts added to each mf_density in density column
}
\description{
calculates normalized mosaic fraction density curves given previously calculated densities
}
