% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{plot_averaged_curves}
\alias{plot_averaged_curves}
\title{plots averaged curves in a nested density object grouped, colored, and faceted by specified variables}
\usage{
plot_averaged_curves(
  nested_density,
  group_str = NULL,
  color_str = NULL,
  facet_str = NULL,
  color_values = NULL
)
}
\arguments{
\item{nested_density}{nested_density object NOT WORKING}

\item{color_values}{manually specified color scale to plot curves by if desired}

\item{group}{variable to group by}

\item{color}{variable to color by}

\item{facet}{variable to facet by}
}
\description{
plots averaged curves in a nested density object grouped, colored, and faceted by specified variables
}
