% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MF_Functions.R
\name{plot_id_sample_heatmap}
\alias{plot_id_sample_heatmap}
\title{plots heatmap depicting binary detection of each ID within each sample}
\usage{
plot_id_sample_heatmap(raw, sample_annotation)
}
\arguments{
\item{raw}{raw object with ID and sample columns}

\item{sample_annotation}{vector of variables by which to annotate heatmap rows (samples)}
}
\description{
plots heatmap depicting binary detection of each ID within each sample
}
